backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
uvicorn[standard]==0.24.0
sqlalchemy==1.4.41
pydantic==1.10.12
